+++ src/Products/urban/migration/migrate_to_250.py
@@ -243,6 +243,26 @@ def migrate_inquiry_parcels(context):
     logger.info("migration step done!")
 
 
+def migrate_remove_prov_in_folderroadtypes(context):
+    logger = logging.getLogger('migrate remove prov in folderroadtypes voc')
+    logger.info("starting migration step")
+    urban_tool = api.portal.get_tool('portal_urban')
+    for folderroadtype in urban_tool.folderroadtypes.objectValues():
+        if folderroadtype.id == "prov":
+            api.content.transition(obj=folderroadtype, to_state='disabled')
+    logger.info("migration step done!")
+
+
+def migrate_disable_natura2000_folderzone(context):
+    logger = logging.getLogger('migrate disable natura2000 folderzone')
+    logger.info("starting migration step")
+    urban_tool = api.portal.get_tool('portal_urban')
+    for folderzone in urban_tool.folderzones.objectValues():
+        if folderzone.id == "znatura2000":
+            api.content.transition(obj=folderzone, to_state='disabled')
+    logger.info("migration step done!")
+
+
 def migrate_inquiry_investigationStart_date(context):
     """
     investigationStart and investigationEnd are no longer optional fields
@@ -279,6 +299,8 @@ def migrate(context):
     migrate_parcellings_folder_allowed_type(context)
     migrate_default_states_to_close_all_events(context)
     migrate_inquiry_parcels(context)
+    migrate_remove_prov_in_folderroadtypes(context)
+    migrate_disable_natura2000_folderzone(context)
     catalog = api.portal.get_tool('portal_catalog')
     catalog.clearFindAndRebuild()
     logger.info("migration done!")